+++ src/templates/lib/adapter-config.d.ts.ts
@@ -2,7 +2,7 @@ import { TemplateFunction } from "../../lib/createAdapter";
 
 const templateFunction: TemplateFunction = answers => {
 
-	const isAdapter = answers.features.indexOf("Adapter") > -1;
+	const isAdapter = answers.features.indexOf("adapter") > -1;
 	if (!isAdapter) return;
 
 	const useTypeScript = answers.language === "TypeScript";